View tests added to test the login authentication and logout token blacklisting.
View tests (controller for the MVT architecture) are located under backend/test/views. Current environment set up should be sufficient to run the Authentication tests. **Test will yield a warning that is not of concern for this commit, but will be addressed in future commits

cd backend
pytest test/views/AuthenticationTest.py