Add GenreTest model validation tests
Note that the object must be created with objects.create, or it doesn't store in the database and is just a Python object.

nullcontext() ensures that the following lines do NOT raise an exception, and fails if it does.

full_clean() re-validates the object, so it will raise a ValidationError if the object's values violate the model validations