Rename foreign keys and squash migrations
Standard practice is for foreign keys to be named the name of the table, as we're not directly referencing the ID, but the object itself.

Generate new migration for this and squash old migrations.

Squashed migrations preserve migration history, but take up less space, and be applied on a fresh database.

Note that if someone is missing at least one migration (but not all), applying migrations will fail, as Django checks migration history and notices a migration is missing. This is why the new migration is not included in the squashed migrations.

Migrations and squashed migrations can be squashed multiple times with

`python manage.py squashmigrations pages {first_migration_name} {last_migration_name}`