Use fixtures and explicitly call model validation
Django doesn't automatically validate an object when saving it, it only enforces database level validations like null=true and unique=True. Other validators (blank=true, minLength) are not enforced.

Since we want to validate that the database is not adding empty names, we must explicitly validate the model

Also note that in the uniqueness test for users, simply putting the fixture in the function definition will populate the database with the user.