Business model validation testing
Note that I use full_clean() to test field-level validations, but .save() and .create() actually saves to the database.

Testing saving to the database needs to be done for cascading foreign keys, null values, etc.

See the field max length test and how it uses the error message to make sure that BOTH fields failed the validation