Add post model and tests
TextField does not have a maximum text limit, so add a validation to cap at 500 characters.
For images, use JSONField to store URLs that will correspond to their URL in the S3 bucket.
Tagged_users is a many-to-many relation so that a post can have multiple tagged users, but only one post owner.