Setup forgot password reset via email
Use HTML to generate email body. Emails only render HTML with inline CSS. The email format should be redesigned in the future.

The request contains an encoded UID and generated token that expires if not used. The frontend gets this info from the URL and includes it in the request to the backend.

Note that there are no visible validations for if there is no user registered with an email, for security purposes. There is an extra timeout included in requests that error, as these usually return immediately, as opposed to requests that actually read a user from the database.