Creation and querying of posts interfaces with S3
We will change post images to be through a public s3 bucket, but this is over private s3 for now. Private s3 will be used for contracts and messages between users.

For private buckets, we need to set up our AWS profile and then make a request to the s3 bucket using that profile. To do this, I used Django's built-in method for generating a pre-signed url for the s3 image with the correct profile. Instead of storing the s3_url in the db, we are opting to create the url using the file_key (stored in the db). This is done for security and storage - the generated links will also expire.

We can still serialize this to include it in the response by using SerializerMethodField.

For public s3 once we switch, we won't need to sign the url.

It seems that generating the presigned link when using the us-east-2 region defaults to not including it in the url, both are accepted domains in next.config.ts for safety.

At the moment, this is only designed for a single png.