Load posts into feed
Toolbars are on most pages - structure of some pages are weird. Currently toolbar is over the text on Discover page, going to transition it to MUI due to lack of CSS experience

MUI suggests using CardActionArea for large onClick events like for the posts. CardActions is used for the like + comment buttons

I don’t want to serialize post comments in side the post serializer because then it sends all comment data. I’d rather paginate the comments separately

Best practices suggest to not use index (which may not map correctly with deletions/adds) and instead use a unique key for an html attribute when mapping

Currently does not query correctly - waiting for Kenna's insight on getting the user from useAuth()