Upload multiple images at once
Posts are saved, serialized, and queried with correct amount of images. Working on saving state and changing between images.

PostForm was a major change that needed to be altered to allow for multiple files. Much of this code is updated from the Django documentation for uploading multiple files, and the clean_file validation is relatively similar.