Upload files to messages + fix image size
Locking box width to a predetermined image width so that all images are the same size, and the same size as the box

Decided to use 'fill' which distorts image, instead of cropping. Can change to crop instead ('cover'), though still would recommend one of the strategies that enforces consistent size and no whitespace.